Create coronavirus.java
An integer matrix of size R x C containing only the values 0,1 and 2 is given as the input to the program. The value 0 indicates an empty space, the value 1 indicates a person is healthy and the value 2 indicates a person is infected by corona virus. Every day the virus is spread from infected person to the other persons. The program must print the minimum number of days required to spread the coronavirus to all individuals. If all the persons can not be affected by the corona virus, the program must print -1 as the output.